Major clean up
