--- requirements.txt
@@ -1,13 +1,0 @@
-certifi==2017.11.5
-chardet==3.0.4
-future==0.16.0
-idna==2.6
-numpy==1.13.3
-OpenNMT-py==0.1
-PyYAML==3.12
-requests==2.18.4
-six==1.11.0
-torch==0.3.0.post4
-torchtext==0.1.1
-tqdm==4.19.4
-urllib3==1.22
